+++ pioreactor/automations/dosing/chemostat.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 
 from pioreactor.automations import events
 from pioreactor.automations.dosing.base import DosingAutomationJob
+from pioreactor.exc import CalibrationError
 from pioreactor.utils import local_persistant_storage
 
 
@@ -22,9 +23,9 @@ class Chemostat(DosingAutomationJob):
 
         with local_persistant_storage("pump_calibration") as cache:
             if "media_ml_calibration" not in cache:
-                raise RuntimeError("Media pump calibration must be performed first.")
+                raise CalibrationError("Media pump calibration must be performed first.")
             elif "waste_ml_calibration" not in cache:
-                raise RuntimeError("Waste pump calibration must be performed first.")
+                raise CalibrationError("Waste pump calibration must be performed first.")
 
         self.volume = float(volume)