+++ pioreactor/automations/dosing/pid_turbidostat.py
@@ -6,6 +6,7 @@ from contextlib import suppress
 from pioreactor.automations import events
 from pioreactor.automations.dosing.base import DosingAutomationJob
 from pioreactor.config import config
+from pioreactor.exc import CalibrationError
 from pioreactor.utils import local_persistant_storage
 from pioreactor.utils.streaming_calculations import PID
 
@@ -28,9 +29,9 @@ class PIDTurbidostat(DosingAutomationJob):
 
         with local_persistant_storage("pump_calibration") as cache:
             if "media_ml_calibration" not in cache:
-                raise RuntimeError("Media pump calibration must be performed first.")
+                raise CalibrationError("Media pump calibration must be performed first.")
             elif "waste_ml_calibration" not in cache:
-                raise RuntimeError("Waste pump calibration must be performed first.")
+                raise CalibrationError("Waste pump calibration must be performed first.")
 
         self.set_target_od(target_od)
         self.volume_to_cycle = None